@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-gem 'rails', '3.2.14'
+gem 'rails', '4.0.0'
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
 gem 'newrelic_rpm'
@@ -10,8 +10,8 @@ gem 'stripe'
 gem 'rails_autolink'
 
 group :assets do
-  gem 'sass-rails', '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails', '~> 4.0.0'
+  gem 'coffee-rails', '~> 4.0.0'
   gem 'uglifier'
 end
 
@@ -22,33 +22,25 @@ gem 'sass'
 gem 'chronic'
 gem 'timelord', '0.0.6'
 gem 'acts_as_list'
-gem 'devise_invitable'
+gem 'devise_invitable', github: 'scambra/devise_invitable'
 gem 'hoptoad_notifier'
-
-group :development do
-  gem 'hpricot'
-  gem 'ruby_parser'
-end
 
 group :development, :test do
   gem 'rspec-rails'
-  platforms :ruby_19 do
-    gem 'ruby-debug19', :require => 'ruby-debug'
-  end
-  platforms :ruby_18 do
-    gem 'ruby-debug'
-  end
+  gem 'pry'
 end
 
 group :test do
   gem 'capybara'
   gem 'capybara-webkit'
+  gem 'selenium-webdriver'
   gem 'timecop'
   gem 'database_cleaner'
-  gem 'cucumber-rails', '1.1.1', require: false
+  gem 'cucumber-rails', require: false
   gem 'bourne'
   gem 'launchy'    # So you can do Then show me the page
-  gem 'factory_girl_rails', '~> 3.3.0'
+  gem 'factory_girl', '~> 3.6.2'
+  gem 'factory_girl_rails', '~> 3.6.0'
   gem 'shoulda', "3.0.0.beta2"
   gem 'email_spec'
   gem 'fakeweb'
