@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Pinchlist::Application.config.secret_token = '9c96dad2b812fa0674a1402b7882c2a0c362f80069843075cb6c072bfe3f3c4c53a1a6ab723cd424f04ef2f683a7f8ca07601e75ab670118913057f7fd5bf114'
+Pinchlist::Application.config.secret_key_base = '9c96dad2b812fa0674a1402b7882c2a0c362f80069843075cb6c072bfe3f3c4c53a1a6ab723cd424f04ef2f683a7f8ca07601e75ab670118913057f7fd5bf114'
